Mini fix for run script

+++ multiscale_benchmark/2022_09_hackathon/Biodynamo/unit_test_chemotaxis/run.sh
@@ -32,4 +32,4 @@ bdm run
 
 # Move the results to the correct folder
 mv output/chemotaxis/metadata $DIR/results
-mv output/position.csv $DIR/results
+[ -f output/position.csv ] && mv output/position.csv $DIR/results
